Refactoring axial expansion
