+++ armi/reactor/converters/axialExpansionChanger/__init__.py
@@ -1,0 +1,29 @@
+# Copyright 2024 TerraPower, LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""Enable component-wise axial expansion for assemblies and/or a reactor."""
+
+# ruff: noqa: F401
+from armi.reactor.converters.axialExpansionChanger.assemblyAxialLinkage import (
+    AssemblyAxialLinkage,
+)
+from armi.reactor.converters.axialExpansionChanger.axialExpansionChanger import (
+    AxialExpansionChanger,
+)
+from armi.reactor.converters.axialExpansionChanger.axialExpansionChanger import (
+    makeAssemsAbleToSnapToUniformMesh,
+)
+from armi.reactor.converters.axialExpansionChanger.expansionData import ExpansionData
+from armi.reactor.converters.axialExpansionChanger.expansionData import (
+    getSolidComponents,
+)
